+++ api/error_messages.go
@@ -25,13 +25,13 @@ const (
 	errNoAccountsFound                 = "no accounts found for address"
 	errNoAssetsFound                   = "no assets found for asset-id"
 	errNoTransactionFound              = "no transaction found for transaction id"
-	errMultipleTransactions            = "multiple transactions found for this txid, please contact us this shouldn't happen"
-	errMultipleAccounts                = "multiple accounts found for this address, please contact us this shouldn't happen"
-	errMultipleAssets                  = "multiple assets found for this id, please contact us this shouldn't happen"
-	errMultipleApplications            = "multiple applications found for this id, please contact us this shouldn't happen"
+	errMultipleTransactions            = "multiple transactions found for this txid, please contact us, this shouldn't happen"
+	errMultipleAccounts                = "multiple accounts found for this address, please contact us, this shouldn't happen"
+	errMultipleAssets                  = "multiple assets found for this id, please contact us, this shouldn't happen"
+	errMultipleApplications            = "multiple applications found for this id, please contact us, this shouldn't happen"
 	errMultiAcctRewind                 = "multiple accounts rewind is not supported by this server"
 	errRewindingAccount                = "error while rewinding account"
-	errLookingUpBlock                  = "error while looking up block for round"
+	errLookingUpBlockForRound          = "error while looking up block for round"
 	errTransactionSearch               = "error while searching for transaction"
 	errZeroAddressCloseRemainderToRole = "searching transactions by zero address with close address role is not supported"
 	errZeroAddressAssetSenderRole      = "searching transactions by zero address with asset sender role is not supported"
